+++ test/unit/view/drag-handle/util/basic-context.js
@@ -1,0 +1,12 @@
+// @flow
+import {
+  styleContextKey,
+  canLiftContextKey,
+} from '../../../../../src/view/context-keys';
+
+const basicContext = {
+  [styleContextKey]: 'hello',
+  [canLiftContextKey]: () => true,
+};
+
+export default basicContext;
